requirements.txt
pygame==2.1.2
RPi.GPIO==0.7.1
